requirements.txt
asgiref==3.7.2
boto==2.49.0
boto3==1.28.21
botocore==1.31.21
click==8.1.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.8
django-import-export==3.2.0
django-ninja==0.22.2
django-storages==1.13.2
djongo==1.3.6
dnspython==2.4.1
et-xmlfile==1.1.0
h11==0.14.0
jmespath==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==8.3.2
pydantic==1.10.12
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.2.4
tablib==3.5.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
xlrd==2.0.1
xlwt==1.3.0
django-cors-headers==3.14.0
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
PyJWT==1.7.1